backend/requirements.txt
pypdf
cohere
python-dotenv
flask
flair
openai